MediFi-GPT/genai/text_splitter.py
from langchain.text_splitter import RecursiveCharacterTextSplitter

def split_into_chunks(text, chunk_size=1000, overlap=200):
    splitter = RecursiveCharacterTextSplitter(chunk_size=chunk_size, chunk_overlap=overlap)
    return splitter.split_text(text)
