MediFi-GPT/requirements.txt
streamlit
openai
faiss-cpu
sentence-transformers
PyMuPDF
python-dotenv